Use thin to get right Content-Length
When the app runs on WEBrick and Transfer-Encoding is set to "chucked",
Content-Length is not correctly calculated.  Actual content size is
always a bit less than Content-Length.  When the app runs on thin, the
problem does not happen.

Additionally, Heroku uses thin by default.  It's better to use thin also
in development environments.